Fix check_collision

+++ include/rectangle.h
@@ -3,21 +3,21 @@
 
 class Rectangle {
     public:
-        Rectangle(double x = 0.0, double y = 0.0, double w = 1.0, double h = 1.0);
+        Rectangle(int x = 0.0, int y = 0.0, int w = 1.0, int h = 1.0);
 
-        double x() const;
-        double y() const;
-        double w() const;
-        double h() const;
-        double area() const;
+        int x() const;
+        int y() const;
+        int w() const;
+        int h() const;
+        int area() const;
 
         bool intersection(const Rectangle& r) const;
 
     private:
-        double m_x; // coordinate x of the rectangle's CENTER
-        double m_y; // coordinate y of the rectangle's CENTER
-        double m_w; // width of the rectangle
-        double m_h; // height of the rectangle
+        int m_x; // coordinate x of the rectangle's CENTER
+        int m_y; // coordinate y of the rectangle's CENTER
+        int m_w; // width of the rectangle
+        int m_h; // height of the rectangle
 };
 
 #endif
